api_server/requirements.txt
click==8.0.3
Flask==2.0.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pydantic==1.8.2
PyYAML==6.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
